* 10.0:
  commit eruit
  order_id is copy=False
  remove order_company_id from sale_order_line because the field company_id  (related=order_id.company_id already existed in sale.py company_id indexed to make normal user performance bettersimpler unlink with create_multi_from_order_line
  Subscription sales: renewal with days offset and leap year hardening
  base_suspend_security and create_multi_lines
  SQL for multiline

+++ sale_advertising_order/models/sale_advertising.py
@@ -141,6 +141,10 @@ class SaleOrder(models.Model):
     advertising = fields.Boolean('Advertising', default=False)
     max_discount = fields.Integer(compute='_amount_all', track_visibility='always', store=True, string="Maximum Discount")
     display_discount_to_customer = fields.Boolean("Display Discount", default=False)
+    company_id = fields.Many2one('res.company', 'Company',
+                                 default=lambda self: self.env[
+                                     'res.company']._company_default_get(
+                                     'sale.order'), index=True)
 
     @api.model
     def default_get(self, fields):
@@ -315,7 +319,6 @@ class SaleOrder(models.Model):
                         line.page_qty_check_create()
             if not olines == []:
                 list = self.env['sale.order.line.create.multi.lines'].create_multi_from_order_lines(orderlines=olines)
-                self._cr.commit()
                 newlines = self.env['sale.order.line'].browse(list)
                 for newline in newlines:
                     if newline.deadline_check():
@@ -551,8 +554,8 @@ class SaleOrderLine(models.Model):
     order_agency_id = fields.Many2one(related='order_id.advertising_agency', relation='res.partner',
                                           string='Advertising Agency', store=True)
     order_pricelist_id = fields.Many2one(related='order_id.pricelist_id', relation='product.pricelist', string='Pricelist')
-    order_company_id = fields.Many2one(related='order_id.company_id', relation='res.company',
-                                         string='Company')
+    company_id = fields.Many2one(related='order_id.company_id',
+                                 string='Company', store=True, readonly=True, index=True)
     discount_dummy = fields.Float(related='discount', string='Agency Commission (%)', readonly=True )
     price_unit_dummy = fields.Float(related='price_unit', string='Unit Price', readonly=True)
     actual_unit_price = fields.Float(compute='_compute_amount', string='Actual Unit Price', digits=dp.get_precision('Product Price'),
@@ -1048,12 +1051,16 @@ class SaleOrderLine(models.Model):
     @api.multi
     def page_qty_check_update(self):
         self.ensure_one()
+        if not self.product_template_id.page_id:
+            return
         self.page_qty_check_unlink()
         self.page_qty_check_create()
 
     @api.multi
     def page_qty_check_unlink(self):
         self.ensure_one()
+        if not self.product_template_id.page_id:
+            return
         res = self.env['sale.advertising.available'].search([('order_line_id', '=', self.id)])
         if res and len(res) > 0:
             res.unlink()
